+++ samples/configure-electronic-navigational-charts/README.metadata.json
@@ -1,10 +1,10 @@
 {
     "category": "Layers",
     "description": "Display nautical charts per the ENC specification.",
-    "formal_name": "AddEncExchangeSet",
+    "formal_name": "ConfigureElectronicNavigationalCharts",
     "ignore": false,
     "images": [
-        "add-enc-exchange-set.png"
+        "configure-electronic-navigational-charts.png"
     ],
     "keywords": [
         "ENC",
@@ -25,7 +25,7 @@
         "https://www.arcgis.com/home/item.html?id=5028bf3513ff4c38b28822d010a4937c"
     ],
     "provision_to": [],
-    "redirect_from": "",
+    "redirect_from": ["/kotlin/sample-code/add-enc-exchange-set/"],
     "relevant_apis": [
         "EncCell",
         "EncDataset",
@@ -33,10 +33,10 @@
         "EncLayer"
     ],
     "snippets": [
-        "src/main/java/com/esri/arcgismaps/sample/addencexchangeset/components/MapViewModel.kt",
-        "src/main/java/com/esri/arcgismaps/sample/addencexchangeset/DownloadActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/addencexchangeset/MainActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/addencexchangeset/screens/MainScreen.kt"
+        "src/main/java/com/esri/arcgismaps/sample/configureelectronicnavigationalcharts/components/MapViewModel.kt",
+        "src/main/java/com/esri/arcgismaps/sample/configureelectronicnavigationalcharts/DownloadActivity.kt",
+        "src/main/java/com/esri/arcgismaps/sample/configureelectronicnavigationalcharts/MainActivity.kt",
+        "src/main/java/com/esri/arcgismaps/sample/configureelectronicnavigationalcharts/screens/MainScreen.kt"
     ],
-    "title": "Add ENC exchange set"
+    "title": "Configure electronic navigational charts"
 }
