+++ samples/configure-electronic-navigational-charts/src/main/AndroidManifest.xml
@@ -6,13 +6,13 @@
     <application><activity
         android:name=".DownloadActivity"
         android:exported="true"
-        android:label="@string/add_enc_exchange_set_app_name">
+        android:label="@string/configure_electronic_navigational_charts_app_name">
 
     </activity>
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/add_enc_exchange_set_app_name">
+            android:label="@string/configure_electronic_navigational_charts_app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
             </intent-filter>
