+++ samples/configure-electronic-navigational-charts/src/main/java/com/esri/arcgismaps/sample/configureelectronicnavigationalcharts/screens/MainScreen.kt
@@ -14,7 +14,7 @@
  *
  */
 
-package com.esri.arcgismaps.sample.addencexchangeset.screens
+package com.esri.arcgismaps.sample.configureelectronicnavigationalcharts.screens
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -23,7 +23,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.arcgismaps.toolkit.geoviewcompose.MapView
-import com.esri.arcgismaps.sample.addencexchangeset.components.MapViewModel
+import com.esri.arcgismaps.sample.configureelectronicnavigationalcharts.components.MapViewModel
 import com.esri.arcgismaps.sample.sampleslib.components.MessageDialog
 import com.esri.arcgismaps.sample.sampleslib.components.SampleTopAppBar
