+++ samples/configure-electronic-navigational-charts/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="add_enc_exchange_set_app_name">Add ENC exchange set</string>
+    <string name="configure_electronic_navigational_charts_app_name">Configure electronic navigational charts</string>
     <string name="enc_res_dir">/hydrography</string>
     <string name="enc_data_dir">/ExchangeSetwithoutUpdates/ENC_ROOT/CATALOG.031</string>
 </resources>
